Fixing issue 549
