+++ bsi/src/test/java/org/roaringbitmap/bsi/BufferBSITest.java
@@ -1,44 +1,57 @@
-package org.roaringbitmap.bsi;
+ package org.roaringbitmap.bsi;
 
-import org.junit.jupiter.api.*;
-import org.roaringbitmap.RoaringBitmap;
+
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+import org.roaringbitmap.bsi.buffer.ImmutableBitSliceIndex;
+import org.roaringbitmap.bsi.buffer.MutableBitSliceIndex;
+import org.roaringbitmap.buffer.ImmutableRoaringBitmap;
+import org.roaringbitmap.buffer.MutableRoaringBitmap;
 
 import java.io.*;
 import java.nio.ByteBuffer;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
+import java.util.concurrent.ExecutionException;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
 /**
- * RBBsiTestBase
- * created by haihuang@alibaba-inc.com on 2021/6/6
+ * TestBase
+ *
  */
-public class RBBsiTestBase {
+public class BufferBSITest {
+
     private Map<Integer, Integer> testDataSet = new HashMap<>();
 
-    private RoaringBitmapSliceIndex bsi;
+    private MutableBitSliceIndex mBsi;
+
+    private ImmutableBitSliceIndex imBsi;
+
 
     @BeforeEach
     public void setup() {
         IntStream.range(1, 100).forEach(x -> testDataSet.put(x, x));
-        bsi = new RoaringBitmapSliceIndex(1, 99);
+        mBsi = new MutableBitSliceIndex(1, 99);
         testDataSet.forEach((k, v) -> {
-            bsi.setValue(k, v);
+            mBsi.setValue(k, v);
         });
+        imBsi = mBsi.toImmutableBitSliceIndex();
     }
 
     @Test
     public void testSetAndGet() {
         IntStream.range(1, 100).forEach(x -> {
-            Pair<Integer, Boolean> pair = bsi.getValue(x);
+            Pair<Integer, Boolean> pair = mBsi.getValue(x);
             Assertions.assertTrue(pair.getRight());
             Assertions.assertTrue(pair.getKey() == x);
         });
 
         IntStream.range(1, 100).forEach(x -> {
-            Pair<Integer, Boolean> pair = bsi.getValue(x);
+            Pair<Integer, Boolean> pair = imBsi.getValue(x);
             Assertions.assertTrue(pair.getRight());
             Assertions.assertTrue(pair.getKey() == x);
         });
@@ -46,9 +59,9 @@ public class RBBsiTestBase {
 
     @Test
     public void testMerge() {
-        RoaringBitmapSliceIndex bsiA = new RoaringBitmapSliceIndex();
+        MutableBitSliceIndex bsiA = new MutableBitSliceIndex();
         IntStream.range(1, 100).forEach(x -> bsiA.setValue(x, x));
-        RoaringBitmapSliceIndex bsiB = new RoaringBitmapSliceIndex();
+        MutableBitSliceIndex bsiB = new MutableBitSliceIndex();
         IntStream.range(100, 199).forEach(x -> bsiB.setValue(x, x));
         Assertions.assertEquals(bsiA.getExistenceBitmap().getLongCardinality(), 99);
         Assertions.assertEquals(bsiB.getExistenceBitmap().getLongCardinality(), 99);
@@ -63,14 +76,14 @@ public class RBBsiTestBase {
 
     @Test
     public void testClone() {
-        RoaringBitmapSliceIndex bsi = new RoaringBitmapSliceIndex(1, 99);
+        MutableBitSliceIndex bsi = new MutableBitSliceIndex(1, 99);
         List<Pair<Integer, Integer>> collect = testDataSet.entrySet()
                 .stream().map(x -> Pair.newPair(x.getKey(), x.getValue())).collect(Collectors.toList());
 
         bsi.setValues(collect, 99, 1);
 
         Assertions.assertEquals(bsi.getExistenceBitmap().getLongCardinality(), 99);
-        final RoaringBitmapSliceIndex clone = bsi.clone();
+        final MutableBitSliceIndex clone = bsi.clone();
 
         IntStream.range(1, 100).forEach(x -> {
             Pair<Integer, Boolean> bsiValue = clone.getValue(x);
@@ -82,9 +95,9 @@ public class RBBsiTestBase {
 
     @Test
     public void testAdd() {
-        RoaringBitmapSliceIndex bsiA = new RoaringBitmapSliceIndex();
+        MutableBitSliceIndex bsiA = new MutableBitSliceIndex();
         IntStream.range(1, 100).forEach(x -> bsiA.setValue(x, x));
-        RoaringBitmapSliceIndex bsiB = new RoaringBitmapSliceIndex();
+        MutableBitSliceIndex bsiB = new MutableBitSliceIndex();
         IntStream.range(1, 120).forEach(x -> bsiB.setValue(x, x));
 
         bsiA.add(bsiB);
@@ -104,14 +117,14 @@ public class RBBsiTestBase {
 
     @Test
     public void TestIO4Stream() throws IOException {
-        RoaringBitmapSliceIndex bsi = new RoaringBitmapSliceIndex(1, 99);
+        MutableBitSliceIndex bsi = new MutableBitSliceIndex(1, 99);
         IntStream.range(1, 100).forEach(x -> bsi.setValue(x, x));
         ByteArrayOutputStream bos = new ByteArrayOutputStream();
         DataOutputStream bdo = new DataOutputStream(bos);
         bsi.serialize(bdo);
         byte[] data = bos.toByteArray();
 
-        RoaringBitmapSliceIndex newBsi = new RoaringBitmapSliceIndex();
+        MutableBitSliceIndex newBsi = new MutableBitSliceIndex();
 
         ByteArrayInputStream bis = new ByteArrayInputStream(data);
         DataInputStream bdi = new DataInputStream(bis);
@@ -128,13 +141,13 @@ public class RBBsiTestBase {
 
     @Test
     public void testIO4Buffer() throws IOException {
-        RoaringBitmapSliceIndex bsi = new RoaringBitmapSliceIndex(1, 99);
+        MutableBitSliceIndex bsi = new MutableBitSliceIndex(1, 99);
         IntStream.range(1, 100).forEach(x -> bsi.setValue(x, x));
         ByteBuffer buffer = ByteBuffer.allocate(bsi.serializedSizeInBytes());
         bsi.serialize(buffer);
 
         byte[] data = buffer.array();
-        RoaringBitmapSliceIndex newBsi = new RoaringBitmapSliceIndex();
+        MutableBitSliceIndex newBsi = new MutableBitSliceIndex();
         newBsi.deserialize(ByteBuffer.wrap(data));
         Assertions.assertEquals(newBsi.getExistenceBitmap().getLongCardinality(), 99);
 
@@ -148,20 +161,44 @@ public class RBBsiTestBase {
 
     @Test
     public void testIOFromExternal() {
-        RoaringBitmapSliceIndex bsi = new RoaringBitmapSliceIndex(1, 99);
+        MutableBitSliceIndex bsi = new MutableBitSliceIndex(1, 99);
         IntStream.range(1, 100).forEach(x -> bsi.setValue(x, x));
 
+
+        ImmutableBitSliceIndex iBsi = bsi.toImmutableBitSliceIndex();
+
         IntStream.range(1, 100).forEach(x -> {
-            Pair<Integer, Boolean> bsiValue = bsi.getValue(x);
+            Pair<Integer, Boolean> bsiValue = iBsi.getValue(x);
             Assertions.assertTrue(bsiValue.getRight());
             Assertions.assertEquals((int) bsiValue.getKey(), x);
         });
     }
 
+    // non parallel operation test
+    @Test
+    public void testSum() {
+        MutableBitSliceIndex bsi = new MutableBitSliceIndex(1, 99);
+        IntStream.range(1, 100).forEach(x -> bsi.setValue(x, x));
+
+
+        ImmutableBitSliceIndex iBsi = bsi.toImmutableBitSliceIndex();
+
+        MutableRoaringBitmap foundSet = MutableRoaringBitmap.bitmapOf(IntStream.range(1, 51).toArray());
+
+        Pair<Long, Long> sumPair = iBsi.sum(foundSet);
+
+        System.out.println("sum:" + sumPair.toString());
+
+        int sum = IntStream.range(1, 51).sum();
+        long count = IntStream.range(1, 51).count();
+
+        Assertions.assertTrue(sumPair.getLeft().intValue() == sum && sumPair.getRight() == count);
+
+    }
 
     @Test
     public void testEQ() {
-        RoaringBitmapSliceIndex bsi = new RoaringBitmapSliceIndex(1, 99);
+        MutableBitSliceIndex bsi = new MutableBitSliceIndex(1, 99);
         IntStream.range(1, 100).forEach(x -> {
             if (x <= 50) {
                 bsi.setValue(x, 1);
@@ -171,66 +208,110 @@ public class RBBsiTestBase {
 
         });
 
-        RoaringBitmap bitmap = bsi.compare(BitmapSliceIndex.Operation.EQ, 1, 0, null);
+        ImmutableRoaringBitmap bitmap = bsi.toImmutableBitSliceIndex().rangeEQ(null, 1);
         Assertions.assertTrue(bitmap.getLongCardinality() == 50L);
-
+        ImmutableRoaringBitmap bitmap129 = bsi.toImmutableBitSliceIndex().rangeEQ(null, 129);
+        Assertions.assertTrue(bitmap129.getLongCardinality() == 0L);
     }
 
 
     // parallel operation test
 
     @Test
-    public void testGT() {
-        RoaringBitmap result = bsi.compare(BitmapSliceIndex.Operation.GT, 50, 0, null);
+    public void testGT()  {
+        ImmutableRoaringBitmap result = imBsi.compare(BitmapSliceIndex.Operation.GT, 50, 0, null);
         Assertions.assertTrue(result.getLongCardinality() == 49);
         Assertions.assertArrayEquals(IntStream.range(51, 100).toArray(), result.toArray());
     }
 
 
     @Test
-    public void testGE() {
-        RoaringBitmap result = bsi.compare(BitmapSliceIndex.Operation.GE, 50, 0, null);
+    public void testGE()  {
+        ImmutableRoaringBitmap result = imBsi.compare(BitmapSliceIndex.Operation.GE, 50, 0, null);
         Assertions.assertTrue(result.getLongCardinality() == 50);
         Assertions.assertArrayEquals(IntStream.range(50, 100).toArray(), result.toArray());
     }
 
     @Test
-    public void testLT() {
-        RoaringBitmap result = bsi.compare(BitmapSliceIndex.Operation.LT, 50, 0, null);
+    public void testLT()  {
+        ImmutableRoaringBitmap result = imBsi.compare(BitmapSliceIndex.Operation.LT, 50, 0, null);
         Assertions.assertTrue(result.getLongCardinality() == 49);
         Assertions.assertArrayEquals(IntStream.range(1, 50).toArray(), result.toArray());
     }
 
 
     @Test
-    public void testLE() {
-        RoaringBitmap result = bsi.compare(BitmapSliceIndex.Operation.LE, 50, 0, null);
+    public void testLE()  {
+        ImmutableRoaringBitmap result = imBsi.compare(BitmapSliceIndex.Operation.LE, 50, 0, null);
         Assertions.assertTrue(result.getLongCardinality() == 50);
         Assertions.assertArrayEquals(IntStream.range(1, 51).toArray(), result.toArray());
     }
 
     @Test
-    public void testRANGE() {
-        RoaringBitmap result = bsi.compare(BitmapSliceIndex.Operation.RANGE, 10, 20, null);
+    public void testRANGE()  {
+        ImmutableRoaringBitmap result = imBsi.compare(BitmapSliceIndex.Operation.RANGE, 10, 20, null);
         Assertions.assertTrue(result.getLongCardinality() == 11);
         Assertions.assertArrayEquals(IntStream.range(10, 21).toArray(), result.toArray());
     }
 
     @Test
-    public void testSum() {
-        RoaringBitmapSliceIndex bsi = new RoaringBitmapSliceIndex(1, 99);
-        IntStream.range(1, 100).forEach(x -> bsi.setValue(x, x));
+    public void testIN() throws ExecutionException, InterruptedException {
 
-        RoaringBitmap foundSet = RoaringBitmap.bitmapOf(IntStream.range(1, 51).toArray());
+        Set<Integer> values = new HashSet<>();
+        int[] valArr = new int[20];
+        for (int i = 0; i < 20; i++) {
+            values.add(i + 1);
+            valArr[i] = i + 1;
+        }
 
-        Pair<Long, Long> sumPair = bsi.sum(foundSet);
 
-        System.out.println("sum:" + sumPair.toString());
+        ExecutorService pool = Executors.newFixedThreadPool(2);
+        ImmutableRoaringBitmap result = imBsi.parallelIn(2,
+                null, values, pool);
+        Assertions.assertTrue(result.getLongCardinality() == values.size());
+        Assertions.assertArrayEquals(valArr, result.toArray());
+        pool.shutdownNow();
+    }
 
-        int sum = IntStream.range(1, 51).sum();
-        long count = IntStream.range(1, 51).count();
 
-        Assertions.assertTrue(sumPair.getLeft().intValue() == sum && sumPair.getRight() == count);
+    @Test
+    public void testTransposeWithCount() throws ExecutionException, InterruptedException {
+        MutableBitSliceIndex bsi = new MutableBitSliceIndex(1, 99);
+        IntStream.range(1, 100).forEach(x -> {
+            if (x <= 30) {
+                bsi.setValue(x, 1);
+            } else if (x <= 60) {
+                bsi.setValue(x, 2);
+            } else {
+                bsi.setValue(x, 3);
+            }
+
+        });
+
+        ExecutorService pool = Executors.newFixedThreadPool(2);
+        ImmutableBitSliceIndex iBsi = bsi.toImmutableBitSliceIndex();
+
+        final MutableBitSliceIndex result = iBsi.parallelTransposeWithCount(null, 2, pool);
+
+        List<Pair<Integer, Integer>> pairs = result.toPairList();
+        pairs.forEach(System.out::println);
+
+
+        Assertions.assertEquals(30, (int) result.getValue(1).getKey());
+        Assertions.assertEquals(30, (int) result.getValue(2).getKey());
+        Assertions.assertEquals(39, (int) result.getValue(3).getKey());
+
+    }
+
+
+    @Test
+    public void testTopK() {
+        MutableBitSliceIndex bsi = new MutableBitSliceIndex(1, 100001);
+        IntStream.range(1, 100001).forEach(x -> bsi.setValue(x, x));
+        long start = System.currentTimeMillis();
+        MutableRoaringBitmap top = bsi.topK(null, 20);
+        long end = System.currentTimeMillis();
+        System.out.println(top.toString() + " \ntime cost:" + (end - start));
     }
